% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_int.R
\name{freq_int}
\alias{freq_int}
\title{Internal frequentist interaction finction}
\usage{
freq_int(
  name,
  covar,
  coefs,
  vals,
  range_size,
  alpha = 0.05,
  round_res = 3,
  control_fdr = FALSE,
  sig_color = "seagreen3",
  non_sig_color = "chocolate",
  line_color = "black",
  color_mid = "white",
  color_low = "#F05039",
  color_high = "#000066",
  color_values = "grey40",
  color_grid = "black",
  grid_density = 0.01,
  grid_spacing = 0.1,
  crosshatch_non_sig = TRUE
)
}
\arguments{
\item{name}{charachter; names of the variables involved in interaction}

\item{covar}{matrix; covariance of relevant parameters}

\item{coefs}{numeric; vector of coefficient values}

\item{vals}{list; values of the variables}

\item{range_size}{see JNK_lm() or JNK_siena()}

\item{alpha}{see JNK_lm() or JNK_siena()}

\item{round_res}{see JNK_lm() or JNK_siena()}

\item{control_fdr}{see JNK_lm() or JNK_siena()}

\item{sig_color}{see JNK_lm() or JNK_siena()}

\item{non_sig_color}{see JNK_lm() or JNK_siena()}

\item{line_color}{see JNK_lm() or JNK_siena()}

\item{color_mid}{see JNK_lm() or JNK_siena()}

\item{color_low}{see JNK_lm() or JNK_siena()}

\item{color_high}{see JNK_lm() or JNK_siena()}

\item{color_values}{see JNK_lm() or JNK_siena()}

\item{color_grid}{see JNK_lm() or JNK_siena()}

\item{grid_density}{see JNK_lm() or JNK_siena()}

\item{grid_spacing}{see JNK_lm() or JNK_siena()}

\item{crosshatch_non_sig}{see JNK_lm() or JNK_siena()}
}
\value{
Returns tables and plots for interactions
}
\description{
Internal frequentist interaction finction
}
