% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JNK_int.R
\name{JNK_int}
\alias{JNK_int}
\title{Function for creating 2- and 3-way plots and tables based on any frequentist multiplicative model.}
\usage{
JNK_int(
  name,
  covar,
  coefs,
  vals,
  range_size,
  alpha = 0.05,
  round_res = 3,
  control_fdr = FALSE,
  sig_color = "seagreen3",
  non_sig_color = "chocolate",
  line_color = "black",
  color_mid = "#EBCC2A",
  color_low = "#3B9AB2",
  color_high = "#F21A00",
  color_values = "grey40",
  color_grid = "black",
  grid_density = 0.01,
  grid_spacing = 0.1,
  crosshatch_non_sig = TRUE
)
}
\arguments{
\item{name}{character; names of the variables involved in interaction}

\item{covar}{matrix; covariance of relevant parameters only}

\item{coefs}{numeric; vector of coefficient values}

\item{vals}{list; values of the variables}

\item{range_size}{see JNK_lm() or JNK_siena()}

\item{alpha}{see JNK_lm() or JNK_siena()}

\item{round_res}{see JNK_lm() or JNK_siena()}

\item{control_fdr}{see JNK_lm() or JNK_siena()}

\item{sig_color}{see JNK_lm() or JNK_siena()}

\item{non_sig_color}{see JNK_lm() or JNK_siena()}

\item{line_color}{see JNK_lm() or JNK_siena()}

\item{color_mid}{see JNK_lm() or JNK_siena()}

\item{color_low}{see JNK_lm() or JNK_siena()}

\item{color_high}{see JNK_lm() or JNK_siena()}

\item{color_values}{see JNK_lm() or JNK_siena()}

\item{color_grid}{see JNK_lm() or JNK_siena()}

\item{grid_density}{see JNK_lm() or JNK_siena()}

\item{grid_spacing}{see JNK_lm() or JNK_siena()}

\item{crosshatch_non_sig}{see JNK_lm() or JNK_siena()}
}
\value{
Returns tables and plots for interactions
}
\description{
Function for creating 2- and 3-way plots and tables based on any frequentist multiplicative model.
}
