% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JNK_lm.R
\name{JNK_lm}
\alias{JNK_lm}
\title{Johnson-Neyman or Johnson-Neyman-Krause plot for lm or glm outputs}
\usage{
JNK_lm(
  modelOut,
  theta1,
  theta2,
  theta3 = NULL,
  control_fdr = FALSE,
  alpha = 0.05,
  round_res = 3,
  range_size = NULL,
  sig_color = "seagreen3",
  non_sig_color = "chocolate",
  line_color = "black",
  color_mid = "white",
  color_low = "#F05039",
  color_high = "#000066",
  color_grid = "black",
  grid_density = 0.01,
  grid_spacing = 0.1,
  crosshatch_non_sig = TRUE
)
}
\arguments{
\item{modelOut}{lm or glm output}

\item{theta1}{character; name of the first variable involved in the interaction}

\item{theta2}{character; name of the second variable involved in the interaction}

\item{theta3}{character; name of the third variable involved in the interaction}

\item{control_fdr}{logical; should Bonferroni-Holms correction be used}

\item{alpha}{numeric; what is the alpha-level}

\item{round_res}{integer; to which level should results be rounded}

\item{range_size}{integer; number of calculated parameter combinations. Default for 2-way is 1000, default for 3-way is 50. Note that for 3-way \code{range_size^2} values will be calculated.}

\item{sig_color}{character; sets color for significant area in 2-way plot}

\item{non_sig_color}{character; sets color for insignificant area in 2-way plot}

\item{line_color}{character; sets color for line in 2-way plot}

\item{color_mid}{character; sets color for \code{0} value in 3-way plot}

\item{color_low}{character; sets color for negative value in 3-way plot}

\item{color_high}{character; sets color for positive values in 3-way plot}

\item{color_grid}{character; sets color of grid}

\item{grid_density}{numeric; sets density of grid}

\item{grid_spacing}{numeric; sets spacing of grid}

\item{crosshatch_non_sig}{logical; should insignificant area be hashed out? If FALSE, significant area will be hashed instead.}
}
\value{
Returns a list with tables for each of the moderated variables and plots for each.
}
\description{
Johnson-Neyman or Johnson-Neyman-Krause plot for lm or glm outputs
}
\examples{
\dontrun{
dat <- data.frame(y = rnorm(100),
                  x = rnorm(100),
                  z = rnorm(100),
                  w = rnorm(100))

res <- lm(y ~ x * z * w,dat)

x2  <- JN.lm(res,
             theta1 = 'x',
             theta2 = 'z')
x2$plots$z

x3  <- JN.lm(res,
             theta1 = 'x',
             theta2 = 'z',
             theta3 = 'w')
x3$plots$z
}
}
