% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JNSienabayes.R
\name{JNSienaBayes}
\alias{JNSienaBayes}
\title{Create Johnson-Neyman plots for Bayesian SAOMs}
\usage{
JNSienaBayes(
  sbo,
  theta_1,
  theta_2,
  theta_3 = NULL,
  theta_int_12,
  theta_int_13 = NULL,
  theta_int_23 = NULL,
  theta_int_123 = NULL,
  theta_1_vals,
  theta_2_vals,
  theta_3_vals = NULL,
  burnIn = NULL,
  thin = 1,
  thresholds = NULL,
  hyper_only = TRUE,
  round_res = 3,
  noTitle = NULL,
  color_mid = "#EBCC2A",
  color_low = "#3B9AB2",
  color_high = "#F21A00",
  color_values = "grey40",
  color_grid = "black",
  grid_density = 0.01,
  grid_spacing = 0.1,
  save = FALSE,
  folder = NULL
)
}
\arguments{
\item{sbo}{multiSiena; multiSiena() output}

\item{theta_1}{numeric; number of the parameter in the effects object. All numbers ignore rates! see 'sbo$effects[sbo$effects$type != 'rate',]'}

\item{theta_2}{numeric; number of the parameter in the effects object. All numbers ignore rates! see 'sbo$effects[sbo$effects$type != 'rate',]'}

\item{theta_3}{numeric; number of the parameter in the effects object. All numbers ignore rates! see 'sbo$effects[sbo$effects$type != 'rate',]'. Default NULL}

\item{theta_int_12}{numeric; number of the interaction between theta_1 and theta_2 in the effects object. All numbers ignore rates! see 'sbo$effects[sbo$effects$type != 'rate',]'}

\item{theta_int_13}{numeric; number of the interaction between theta_1 and theta_2 in the effects object. All numbers ignore rates! see 'sbo$effects[sbo$effects$type != 'rate',]'. Default NULL}

\item{theta_int_23}{numeric; number of the interaction between theta_1 and theta_2 in the effects object. All numbers ignore rates! see 'sbo$effects[sbo$effects$type != 'rate',]'. Default NULL}

\item{theta_int_123}{numeric; number of the interaction between theta_1, theta_2, and theta_3 in the effects object. All numbers ignore rates! see 'sbo$effects[sbo$effects$type != 'rate',]'. Default NULL}

\item{theta_1_vals}{numeric; possible range of the statistic the parameter is multiplied with.}

\item{theta_2_vals}{numeric; possible range of the statistic the parameter is multiplied with.}

\item{theta_3_vals}{numeric; possible range of the statistic the parameter is multiplied with. Default NULL}

\item{burnIn}{numeric; burn-in iteration to be removed. Default NULL}

\item{thin}{numeric; Should the posteriors be thinned? Default = 1, i.e., every iteration is used. Thin = 2 means every 2nd iteration is used.}

\item{thresholds}{numeric; threshold for 'significance' hashing. (Note that all concerns with regards to 'significance' in Bayesian models apply.) Default = NULL}

\item{hyper_only}{logical; should only the hyper parameter be used or plots be made for the random parameters?}

\item{round_res}{numeric; to how many digits should results be rounded? Default = 3}

\item{noTitle}{character; You can provide a title to be used for the plots. Titles can always be changed later by using regular ggplot2 coding.}

\item{color_mid}{character; Color chosen for the mid-point of the scale. Default = 'white'}

\item{color_low}{character; color chosen for low values. Default = '#F05039', (red)}

\item{color_high}{character; color chosen for high values. Default = '#000066', (blue)}

\item{color_values}{character; color chosen for the numbers. Default = 'grey40'}

\item{color_grid}{character; color chosen for the grid. Default = 'black'}

\item{grid_density}{numeric; density of the hash-grid. Default = 0.01}

\item{grid_spacing}{numeric; spacing of the hash-grid. Default = 0.1}

\item{save}{logical; should the plots be saved directly with ggsave(). This is advisable if you have many plots and want to look at them with more ease. See also the folder parameter below. Default = FALSE}

\item{folder}{character; Where should the plots be saved. If NULL, a new folder for the plots - int3ract JNplots - will be created. If that folder does not exist, RStudio will ask if it should be created.}
}
\value{
list of plots.
}
\description{
Create Johnson-Neyman plots for Bayesian SAOMs
}
