% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.JN}
\alias{print.JN}
\title{Print a Johnson-Neyman analysis}
\usage{
\method{print}{JN}(x, ...)
}
\arguments{
\item{x}{a \code{JN} object.}

\item{...}{ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Johnson-Neyman analysis
}
